+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "~server/*": ["server/*"]
+      "~server/*": ["server/*"],
+      "~utils/*": ["utils/*"]
     },
     "allowJs": false,
     "strict": true,
