+++ utils/time.ts
@@ -1,0 +1,1 @@
+export const dayInSecs = 86400;
